lib: add gitservice
This is a service that was originally written for src-cli, but then copy
pasted into gitserver. It has evolved a bit and occasionally they have
been out of sync. So instead we are moving the service into a new
package for both gitserver and src-cli to use.
